requirements.txt
numpy
pandas
Pillow
scikit-learn
tensorflow
flask
matplotlib
nltk